outsourced active user checkup
